float32

+++ server/server.py
@@ -76,7 +76,7 @@ async def transcribe_stream(ws: WebSocket):
 
             if len(audio_data) > 200000:  # Threshold for triggering transcription
                 await ws.send_json({"transcribing": True})
-                data_input_from_source = np.frombuffer(audio_data, dtype=np.int16).astype(np.float32)
+                data_input_from_source = np.frombuffer(audio_data, dtype=np.float32).astype(np.float32)
                 # data_input = np.concatenate((warm_up_audio, data_input_from_source), axis=0)
                 segments, info = model.transcribe(audio=data_input_from_source, beam_size=5, vad_filter=True)
